python/api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
vtracer==0.6.10
Pillow==10.0.1
requests==2.31.0
Werkzeug==2.3.7
